- fix compiler warning with new libpng
git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@1500 e54a6e83-5905-42d5-8d5c-058d10e6a962

+++ src/driver/pictureviewer/png.cpp
@@ -107,7 +107,7 @@ int fh_png_load(const char *name,unsigned char **buffer,int* /*xp*/,int* /*yp*/)
 	if (width * 3 != png_get_rowbytes(png_ptr, info_ptr))
 	{
 		printf("[png.cpp]: Error processing %s - please report (including image).\n", name);
-		printf("           width: %lu rowbytes: %lu\n", width, png_get_rowbytes(png_ptr, info_ptr));
+		printf("           width: %lu rowbytes: %lu\n", (unsigned long)width, (unsigned long)png_get_rowbytes(png_ptr, info_ptr));
 		fclose(fh);
 		return(FH_ERROR_FORMAT);
 	}
